+++ packages/cluster-workflow/test/utils.ts
@@ -8,7 +8,7 @@ import type { Mock } from "vitest"
 
 export function mockEffect<A, E>(
   impl: () => Exit.Exit<A, E>
-): { effect: Effect.Effect<A, E>; spy: Mock<[], Exit.Exit<A, E>> } {
+): { effect: Effect.Effect<A, E>; spy: Mock<() => Exit.Exit<A, E>> } {
   const spy = vi.fn(impl)
   const effect = pipe(Effect.sync(spy), Effect.flatten)
   return { spy, effect }